cksum.c:123:2: warning: conversion to 'uint32_t' from 'int' may change the sign of the result cksum.c:124:2: warning: conversion to 'uint32_t' from 'int' may change the sign of the result cksum.c:125:2: warning: conversion to 'uint32_t' from 'int' may change the sign of the result

+++ src/support/cksum.c
@@ -120,10 +120,10 @@ __wt_cksum(void *chunk, uint32_t bytes)
 	uint8_t *data;
 
 	data = chunk;
-	result = *data++ << 24;
-	result |= *data++ << 16;
-	result |= *data++ << 8;
-	result |= *data++;
+	result  = (uint32_t)*data++ << 24;
+	result |= (uint32_t)*data++ << 16;
+	result |= (uint32_t)*data++ << 8;
+	result |= (uint32_t)*data++;
 	result = ~result;
 	bytes -= 4;
